requirements.txt
aenum==2.2.1
aiohttp==3.6.0
aiohttp-cors==0.7.0
appdirs==1.4.3
astor==0.8.0
async-timeout==3.0.1
asynctest==0.13.0
attrs==19.1.0
autopep8==1.4.4
base58==1.0.3
bitcoin==1.1.42
black==19.10b0
certifi==2019.6.16
chardet==3.0.4
click==7.1.2
coverage==4.5.4
coveralls==1.8.2
coz-bytecode==0.5.1
docopt==0.6.2
ecdsa==0.13.2
filelock==3.0.12
idna==2.8
logzero==1.7.0
mmh3==2.5.1
mpmath==1.1.0
multidict==4.5.2
neo-python-rpc==0.2.1
pathspec==0.8.0
pycodestyle==2.5.0
pymitter==0.2.3
regex==2020.4.4
requests==2.22.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.3
yarl==1.3.0

